corrections of the 21 sprint review
