+++ app/src/main/java/com/example/playlistmaker/domain/search/api/repository/TrackRepository.kt
@@ -6,4 +6,10 @@ import kotlinx.coroutines.flow.Flow
 
 interface TrackRepository {
     fun search(term: String): Flow<Resource<List<Track>>>
+
+    suspend fun save(track: Track)
+
+    suspend fun delete(track: Track)
+
+    fun getTracks(): Flow<List<Track>>
 }
